+++ src/player/GameObjectFactory.h
@@ -10,11 +10,14 @@
 --	DATE:
 --					March 9, 2012
 --	REVISIONS:
+--					March 14, 2012 (Roger)
+--						- added type casts for double values
 --
 --	DESIGNER:
 --					Zach Smoroden
 --	PROGRAMMER:
 --					Zach Smoroden
+--					Roger Fan
 --	NOTES:
 ----------------------------------------------------------------------*/
 #ifndef GAMEOBJECTMOVEABLE_H_
@@ -41,25 +44,25 @@ class GameObjectFactory
 			if (type == "P")
 			{
 				*in_ >> objID >> degree >> posX >> posY >> playerID >> speed >> ttl >> damage;
-				return new GOM_Projectile(objId, degree, posX, posY, playerID, speed, ttl, damage);
+				return new GOM_Projectile(objId, double(degree), double(posX), double(posY), playerID, speed, ttl, damage);
 			}
 			//Ship
 			if (type == "S")
 			{
 				*in_ >> objID >> degree >> posX >> posY >> playerID >> speed >> health >> attackPower;
-				return new GOM_Ship(objId, degree, posX, posY, playerID, speed, health, attackPower);
+				return new GOM_Ship(objId, double(degree), double(posX), double(posY), playerID, speed, health, attackPower);
 			}
 			//Obstacle  
 			if (type == "O")
 			{
                 *in_ >> objID >> type >> degree >> posX >> posY;
-				return new GOS_Obstacle(objId, type, degree, posX, posY);
+				return new GOS_Obstacle(objId, type, double(degree), double(posX), double(posY));
 			}
 			//Powerup  
 			if (type == "B")
 			{
                 *in_ >> objID >> type >> degree >> posX >> posY;
-				return new GOS_Powerup(objId, type, degree, posX, posY);
+				return new GOS_Powerup(objId, type, double(degree), double(posX), double(posY));
 			}
 			
 			return 0;  // if it's not one of the valid types
